+++ src/main/java/nl/enjarai/showmeyourskin/mixin/PlayerEntityModelMixin.java
@@ -2,8 +2,10 @@ package nl.enjarai.showmeyourskin.mixin;
 
 import net.minecraft.client.model.ModelPart;
 import net.minecraft.client.render.entity.model.PlayerEntityModel;
+import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.Items;
 import nl.enjarai.showmeyourskin.config.ModConfig;
 import nl.enjarai.showmeyourskin.util.CombatLogger;
 import org.spongepowered.asm.mixin.Final;
@@ -22,8 +24,9 @@ public abstract class PlayerEntityModelMixin<T extends LivingEntity> {
             at = @At(value = "TAIL")
     )
     private void fixCapeAngle(T livingEntity, float f, float g, float h, float i, float j, CallbackInfo ci) {
-        if (livingEntity instanceof PlayerEntity &&
-                !ModConfig.INSTANCE.getApplicable(livingEntity.getUuid()).shouldTransformCape() &&
+        var applicable = ModConfig.INSTANCE.getApplicable(livingEntity.getUuid());
+        if (livingEntity instanceof PlayerEntity player &&
+                !applicable.shouldTransformCape(player) &&
                 !CombatLogger.INSTANCE.isInCombat(livingEntity.getUuid())) {
             if (livingEntity.isInSneakingPose()) {
                 cloak.pivotZ = 1.4F;